pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grove-kb"
version = "0.1.0"
description = "LLM-compiled knowledge bases. Raw documents in, structured wiki out."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Simon Allen", email = "simon@digitalwonderlab.com" },
]
keywords = ["knowledge-base", "llm", "wiki", "obsidian", "second-brain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "anthropic>=0.25",
    "httpx>=0.27",
    "gitpython>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "sqlite-utils>=3.36",
]

[project.optional-dependencies]
full = [
    "pymupdf4llm>=0.0.17",
    "pdfminer.six>=20221105",
    "readability-lxml>=0.8",
    "markdownify>=0.13",
    "sqlite-vec>=0.1",
    "ollama>=0.2",
    "fastapi>=0.115",
    "uvicorn>=0.34",
]
azure = [
    "azure-ai-inference>=1.0.0b7",
    "azure-identity>=1.19",
]
openai = [
    "openai>=1.50",
]
dev = [
    "ruff",
    "black",
    "pytest",
    "pytest-asyncio",
    "mypy",
    "httpx",
]

[project.scripts]
grove = "grove.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["grove"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]

[tool.black]
target-version = ["py311"]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.11"
strict = true
